feat(template): fold macros, drop Joern's stub debris, quiet the CLI
Three changes, all in what Joern hands us for one translation unit.

Macros. Joern runs no preprocessor: it models `#define X 512` as an external
method and every *use* as a call with the expansion hung underneath. The
converter folds a use into that expansion now, so a macro bound reads as the
number it stands for and a macro-wrapped `strcpy` reaches the sink tables under
its real name. `--no-replace-macro` keeps the old shape and reaches the five
stages that build a Template, instead of sitting unused on `cpg`. The
pseudo-call detection in ast/ and dfg/ stays, for templates written to disk
before this. `ctx_upper_bound_norm` and `buffer_size_norm` are fractions the
extractor divides into but were declared `int`, so the first bound that ever
resolved failed validation instead of being reported.

Stub debris. Joern stubs every function it never sees a body for -- library
calls, the `<operator>.*` family, and macros -- as a METHOD holding only
`p1..pn` parameters typed ANY and an empty block. The converter's passthrough
dropped the METHOD but promoted its children, so a 15-line file produced 52
roots, all but one of them debris: beside the TranslationUnit for library and
operator stubs, which hang off a detached NAMESPACE_BLOCK, and inside it for a
macro, which hangs off the file. They are discarded at conversion now. Nothing
real goes with them -- every fixture's AST and DFG is byte-identical, and
`template_root_count` is the only field that moves in the goldens.

CLI noise. A five-file run printed 514 lines over its own progress bar.
`_quiet_logging` had never once worked: it reached for a logback class, but
Joern binds SLF4J to Log4j2, so every call threw into its own bare `except`.
Joern reads its root level from SL_LOGGING_LEVEL, set to warn before the JVM
starts. Its progress banners are plain println that no log level touches, so
the JVM's stdout and stderr go to a buffer -- installed once, because Log4j2's
console appender binds to whatever stream is current when it first configures
-- replayed under `--debug` and attached to the error when a file fails. Two
JVM flags silence the startup warnings that were landing inside the bar itself.
Messages go through `tqdm.write`: a single per-file error used to close the bar
and drop the reference, removing it for the rest of the run.

Quieting Joern exposed a segfault in JPype's atexit hook, which its own source
warns can crash when a Java thread is waiting for the GIL. The suite passed and
then died with signal 11; `destroy_jvm = False` skips the blocking
DestroyJavaVM call and it exits 0 again.

+++ packages/ssat/src/ssat/ast/validate.py
@@ -15,10 +15,14 @@ class IASTFeatureModel(BaseModel):
     in_loop: int
     is_loop: int
     ctx_guard_strength: int
-    ctx_upper_bound_norm: int
+    # Fractions, not counts: the extractor divides. Declared `int` until a bound
+    # actually resolved, which never happened while every macro comparison read
+    # as unbounded -- 0.0 coerces to an int and 1/256 does not, so the first
+    # resolvable bound failed validation here instead of being reported.
+    ctx_upper_bound_norm: float
     is_buffer_decl: int
     buffer_size_state: int
-    buffer_size_norm: int
+    buffer_size_norm: float
     call_sem_cat_id: int
     call_flag_danger_unbounded: int
     call_flag_len_linked_to_dst: int
